fix i200706 show options in alphabetical order
Previously showed in hashed value order. This made finding the
right folder or config file difficult.
Shows these and the selection values in alphabetical order.
Treating that as a breaking change, in case someone had memorized
the previous menu values.
Does not require typing the preceeding zero, though it will display
it so.
Provides more example folders to test this functionality against.