fix i200706 zero pad transition options
Ensures that options that transition to some other selection
list also control the decision whether to zero pad the currently
relevant options. This had previously seen that less than 9 options
for a config were available, but with the four transion options,
this surpassed the threshold and exposed the user to lexical
sorting again. Increments the compatible version, as the selection
input doesn't change and a person would have seen the same number
previously, albeit in the wrong order.